Moved load and clean functions to a separate .py file
loadBFRresults.py was created
